Rank teams in the qualifying matches of an event by a scoring variable

All teams are ranked (if possible) by their contribution to the scoring variable.

Usage

get_ranking_by(data, variable, method = "normal")

Arguments

data

tibble of match information - it is assumed that each line corresponds to the score breakdown of one team in one event's match.

variable

numeric measurement.

method

character which model should be calculated? Either normal or loglinear.

Value

tibble of the ranked teams and the number of matches (n) that the ranking is based on.

Details

In order to model each team's contribution, a wide data set has to be constructed, with each team as a variable, and each row as one of the alliances in one match and their score. An alliance in a match consists of exactly three robots. This requirement is checked before modelling and if not fulfilled, results in a warning. A normal model is then fitted, or, if method = "loglinear" a log-linear model is fitted, if all values in the scoring variable are non-negative. The coefficients of this model can be interpreted as the average (log) contribution of that team to the score of a match. The OPR (offensive power rating) corresponds to a call to get_ranking_by with score as the variable.

Examples

# example code
# get all matches for one event:
clash_matches <- get_records("event/2024cttd/matches")
# get the score breakdown for each of the teams in each match
clash_details <- get_match_details(clash_matches)

library(dplyr, quietly=TRUE)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# now get the contribution to the score:
clash_details %>% filter(comp_level == "qm") %>% # get the qualifying matches
  get_ranking_by(score)
#> # A tibble: 42 × 3
#>    team_key score     n
#>    <chr>    <dbl> <int>
#>  1 frc3284   44.6     8
#>  2 frc5126   35.6     8
#>  3 frc1108   31.6     8
#>  4 frc1987   30.7     8
#>  5 frc3928   28.1     8
#>  6 frc1986   26.8     8
#>  7 frc2357   26.8     8
#>  8 frc6424   24.9     8
#>  9 frc1769   23.9     8
#> 10 frc5801   23.9     8
#> # ℹ 32 more rows
# These coefficients correspond to the OPR 

# team contribution to the score without counting the opponents' fouls:
clash_details %>% filter(comp_level == "qm") %>% # get the qualifying matches
  get_ranking_by(score-foulPoints)
#> # A tibble: 42 × 3
#>    team_key `score - foulPoints`     n
#>    <chr>                   <dbl> <int>
#>  1 frc1108                  32.4     8
#>  2 frc3284                  30.1     8
#>  3 frc1986                  28.7     8
#>  4 frc6424                  28.7     8
#>  5 frc1987                  27.5     8
#>  6 frc3928                  24.1     8
#>  7 frc2357                  23.1     8
#>  8 frc5126                  22.9     8
#>  9 frc5801                  21.0     8
#> 10 frc935                   20.3     8
#> # ℹ 32 more rows