Rank teams in the qualifying matches of an event by a scoring variable

All teams are ranked (if possible) by their contribution to the scoring variable.

Usage

get_ranking_by(data, variable, method = "normal")

Arguments

data

tibble of match information - it is assumed that each line corresponds to the score breakdown of one team in one event's match.

variable

numeric measurement.

method

character which model should be calculated? Either normal or loglinear.

Value

tibble of the ranked teams and the number of matches (n) that the ranking is based on.

Details

If all values in the scoring variable are positive, a log linear model is fitted otherwise a normal (Gaussian) model is fitted. The OPR

Examples

# example code
# get all matches for one event:
clash_matches <- get_records("event/2024cttd/matches")
# get the score breakdown for each of the teams in each match
clash_details <- get_match_details(clash_matches)

# now get the contribution to the score:
clash_details %>% filter(comp_level == "qm") %>% # get the qualifying matches
  get_ranking_by(score)
#> Error in clash_details %>% filter(comp_level == "qm") %>% get_ranking_by(score): could not find function "%>%"
# OPR is calculated with a normal model of the score without foulpoints
clash_details %>% filter(comp_level == "qm") %>% # get the qualifying matches
  get_ranking_by(score-foulPoints)
#> Error in clash_details %>% filter(comp_level == "qm") %>% get_ranking_by(score -     foulPoints): could not find function "%>%"